Libro busca correctamente, se valida la comunicacion. Libro listo para el ataque >:D

+++ src/main/java/dfs1103/maq/bibiliotecaam/controller/LibroController.java
@@ -46,7 +46,7 @@ public class LibroController {
     }
 
     @GetMapping("/precio/{precio}")
-    public ResponseEntity<List<LibroResponseDTO>> obtenerPorPrecio(Long precio){
+    public ResponseEntity<List<LibroResponseDTO>> obtenerPorPrecio(Integer precio){
         return ResponseEntity.ok(libroService.obtenerPorPrecioMenorQue(precio));
     }
